[BREAKING] rewrite how native browser dialogs are handled
Removed the methods to handle dialogs after-the-fact, instead
you must pre-register a callback to handle any dialogs you expect.
Unexpected dialogs will be closed and then an exception thrown.

This ensures that `wait`, `evaluateScript` etc work as expected in
cases where native dialogs are present, without throwing odd
errors about timeouts on the websocket - which are in fact just
because Chrome has paused execution while it waits for a user
interaction with the dialog.