Fix handling ConnectionException to get the stream metadata array
Since textalk/websocket-php@1449f79fc2b224fe71cf4d06055f41f03ea9b462
the stream metadata is provieded as a property on the exception
itself and is no longer json-encoded within the exception message.

This was causing connection exceptions to have a fatal error
attempting to parse the now-nonexistent-JSON. With luck this will
also allow the existing code in the driver to detect & handle
timeout exceptions properly.