Treat a `null` websocket return as an exception
Previously this was silently skipping out of the waitFor loop,
I don't think that can be right. There's no comment or
information in the code or commit history to indicate why this
was the case - it has been there ever since the very first
commit. Possibly it was originally meant to be a `continue`?

Even if we do encounter this happening, it seems entirely wrong
to allow the calling code to continue without waiting for any
further events or any indication that the thing it was actually
waiting for has happened.

Therefore throw an exception and we can then explore if / when
/ how to deal with it once we know when & why it happens.