Remove duplications between waitForDom & waitForLoad
As it stands they are essentially identical. waitForLoad can't return
until we've seen a frameStoppedLoading, but we don't get that until
after the Page.loadEventFired and by then document.readyState can only
be *complete*.

And, asking if document.readyState===complete runs waitForLoad again
internally, so there's a lot of recursion/looping/hidden waitForLoad that
we don't need.