Fix detecting "page load" after internal navigation
If you click (or do) anything that causes the URL to change
either with e.g. history.replaceState or by clicking a link
to #wherever on the same doc, Chrome sends:
 - Page.frameStartedLoading
 - Page.navigatedWithinDocument
 - Page.frameStoppedLoading

But obviously there is no new Page.loadEventFired. So, for
the moment we need to treat Page.navigatedWithinDocument
as being the equivalent to loadEventFired.

It may actually be better to use Page.frameStoppedLoading
if we always get that, then there's a simple startedLoading
 / stoppedLoading pair?