Apply page load timeout from when page starts loading
If there's an error during waitForLoad (e.g. due to chrome crash
or just brokenness) in the scenario, and then hooks also need
to run things that waitForLoad, it was getting the timeout for
each waitForLoad call.

But actually pageload timeout should just be from the time it
started loading, so subsequent (e.g. ->getCurrentUrl()) get an
almost immediate timeout rather than giving it another
30 seconds.