Don't use the websocket for getting current URL
We ask for the URL e.g. for the SaveFailingPagesListener, and for the
step that records where a scenario started. If the last page never
loaded properly then waitForLoad can wait for yonks before it will
succeed, and it will do this on every scenario which I think is why
they take 28 minutes...