Don't take `Animation` events, they aren't actually used
Animation events produce a flood of events on the dev-tools
protocol particularly (in many projects) around the time of
click, page-load etc events. This may be contributing to
backlogging the socket and the driver occasionally missing
a more important page completion etc event.

The rate of animation events is particularly high because
we are also setting a massive animation playback rate so
animations start and stop very quickly. It is therefore
virtually impossible to actually interact with an element
during an animation.

The code *looks* like it is sleeping when animations begin
- this was introduced in b0239834 and "fixed" in 0b8113f4.
However, in practice that fix actually just disabled the
sleep because the parameter from DevTools is actually in
`$data['params']['animation']['source']['duration']` (note
the extra `animation` in the path). So the `if (!empty(`
has always been returning false and the driver has never
actually been running any code in response to these events.

If there was a desire / need to reintroduce sleeping
during animations in future, we'd need to review the units
of the `duration` value, and whether just multiplying that
by 10 is even close to the actual time taken when playing
back at the forced accelerated rate.