Remove handler for Security.certificateError
DevTools has deprecated the old flow of setting overrideCertificateErrors
and then handling an event notification. Instead they are just ignored
upfront.

The driver was migrated to the new flow in 457160a5 in 2018, but this
(now-dead) code was not removed at the time. These events will not have
been arriving since then, but leaves another place in the code where it
looks like the page state might change.