Switch back to the main mink driver testsuit
At some point, chrome-mink-driver switched to a custom fork of
mink/driver-testsuite which is now a long way behind the actual
one.

https://github.com/adorin/driver-testsuite/compare/master...minkphp:driver-testsuite:master

As far as I can see, all bar one of the changes in the fork are
now in the upstream package, but there are also new tests in
upstream which are not being covered on chrome-mink-driver.

I'm curious to see if these newer tests will flush out the issues
we have seen in actual CI runs, rather than having to add separate
tests for them.