chore(reports): Move coverage test run to report stage
When located in 'test' stage, the most visible test run (top of
alphabetic list) was also the noisiest (prints coverage to output).
This reduced the readability of test output.

Instead, I'll rename this and move it to a report stage. For most test
failures, reading whatever job is at the top of the test stage is fine,
since Gitlab will surface the failing tests to top.

We don't need to capture JUnit output from this run IMO, since that just
inflates the count of tests reported, but doesn't really cover any
additional matrix point - the coverage report uses the default PHP and
Chrome versions, which should already be covered by whatever PHP job is
the current default version.