Don't fake event timestamps when clicking elements
As described in the failing testcase introduced in the previous
commit, the fake timestamp no longer produces a realistic value
in Chrome after version 49. This was causing some client-side
code to incorrectly discard click events triggered through the
driver. The timestamp is not required and will be populated by
Chrome if not present.