Test events fired when calling `setValue` on text-type fields
This is an alternative approach from the one taken in
https://gitlab.com/behat-chrome/chrome-mink-driver/-/merge_requests/134
based on asserting an explicit list of events in sequence.

The event assertions were captured by working through the complete
W3C list of supported input types in Chrome 115.0 and using the test
code to record real keyboard input into all the fields that support
it.

I believe that this provides a conclusive list of how a real Chrome
will behave with all known fields (plus an arbitary `made_up_type`
which is treated as text per the HTML spec).

The current failures and warnings in this testcase match a number of
existing bugs reported on the driver:

* https://gitlab.com/behat-chrome/chrome-mink-driver/-/merge_requests/151
* https://gitlab.com/behat-chrome/chrome-mink-driver/-/issues/139
* https://gitlab.com/behat-chrome/chrome-mink-driver/-/issues/111
* https://gitlab.com/behat-chrome/chrome-mink-driver/-/merge_requests/112
* https://gitlab.com/behat-chrome/chrome-mink-driver/-/issues/95
* https://gitlab.com/behat-chrome/chrome-mink-driver/-/issues/92
* https://gitlab.com/behat-chrome/chrome-mink-driver/-/merge_requests/124