Remove duplicate keypress events triggered during setValue
When piping keys into input elements, the driver is sending a keydown
directly to Chrome, then sending a simulated keydown event through JS.
The same happens for keyup.

It's not clear whether this was ever required - it was introduced by
https://gitlab.com/behat-chrome/chrome-mink-driver/-/merge_requests/56
which states that the keyboard events cannot be captured by javascript.

However the newly-added testcase proves that the devtools commands
*do* trigger key(down|up) events that can be intercepted by javascript.

Without test code at the time, it's not possible to tell whether the
original merge was incorrect (perhaps e.g. due to an unrelated
client-side failure to catch the events) or whether browser behaviour
has changed since 2018.

Either way, it seems clear that these duplicate events are redundant
and risk causing hard-to-debug behavior in client-side code as they
cannot be replicated by user input.