Don't send duplicate keyDown for newlines in setValue
The newline character was being sent first with `chr(13)` and then
as a separate pair of keydown and keyup events. It looks like the
duplicate event was added by accident in 455dd9a5 when fixing
newlines being missing from a textarea.

The code before that commit was already sending a keyDown with a
`'text' => "\n"` which was evidently not affecting the value of
the field, but does fire an additional keyDown event in JS-land.
The fix added a new call to send a `'text' => chr(13)` when in
reality it should just have fixed the argument on the event that
was already being dispatched.