Fix deprecation due to null current_value for contenteditable
setValue was causing deprecation warnings in php >= 8.1 when setting
a value on an empty contenteditable element. This is because
`getValue()` returns `''` for an empty input or textarea, but `null`
for an empty `contenteditable` element.

This was first reported by jidrone and a fix proposed but it was not
clear where the null value was coming from.

I have gone with a simpler null-coalescing operator to reduce the
size of the diff & the requirement for an extra conditional block in
that method.

Closes https://gitlab.com/behat-chrome/chrome-mink-driver/-/merge_requests/151