Switch to phrity/websocket forked websocket library
This is a fork created by the developer who has been the maintainer of
textalk/websocket since v1.3. Per this comment
https://github.com/sirn-se/websocket-php/issues/5#issuecomment-1542182564
he no longer has access to the textalk repositories so forked the project
in order to keep maintaining it.

textalk/websocket has not officially been marked as abandoned, but the
phrity/ version is currently more up to date, and appears to be more
likely to be maintained in future.