Remove test assertion of the TimeoutException message wording
The message itself comes from the php-websocket library, and
the wording depends on the installed version (it has changed
in phrity/websocket:1.7.0).

In the past, chrome-mink-driver had runtime code which depended
on the exact wording of this exception but that was removed in
4365fc4e - the message text is no longer relevant to this
package so there is no need to couple the tests to the internals
of php-websocket.