Remove inoperative timeout/retry handling based on StreamReadException
This was introduced in an attempt to improve handling stream timeouts
and connection failures.

However, it currently does nothing because StreamReadException can
never be thrown with any recent version of php-websocket. The code
attempts to convert a ConnectionException to a StreamReadException
if the exception message contains `{` followed by json-encoded
stream metadata.

But since v1.5.0 of textalk/websocket this has not been the case:
the library removed the JSON from the message in this commit
https://github.com/Textalk/websocket-php/commit/1449f79fc2b224fe71cf4d06055f41f03ea9b462
and instead provides it through a separate `->getData()` method on
the ConnectionException instance.

Additionally, since v1.4.0 the specific exceptions we are looking
for are thrown as a dedicated TimeoutException (subclass of
ConnectionException) so there is no need to convert them to
a custom type.