Deprecate and remove broken `canDevToolsConnectionBeEstablished`
This probably worked when originally introduced. However, the library
itself has only ever called the method on a ChromePage as part of
attempting to decide how to handle a websocket timeout.

This used to go to a hardcoded URL but in 2.8.0 it was changed
to use the `url` injected to the class:
https://gitlab.com/behat-chrome/chrome-mink-driver/-/commit/4dc97d5553df0bcb6cdb9c1375b34ddf8d9b22c3

Unfortunately this overlooked that the URL injected into
ChromePage is the *websocket* URL, not the API base URL.
Therefore it is not possible to simply append the JSON
url path, and this method has been guaranteed to return
false since that change.

It would be possible to fix this, but it is now unused by the
library and there does not seem to be any good reason to
assume that the reachability (or not) of the main browser
JSON API has any bearing on whether a single tab's websocket
that has timed out (but not hit eof) is still valid / healthy.

As it stands, when the library was still throwing StreamReadException,
the `false` returned here was causing a lot of false positive test
failures claiming that the browser had crashed when in fact it
was simply blocked on user input or running slightly slower than
the configured timeout.