fix: Remove duplicate closing non-main windows in `ChromeDriver::close()`
`ChromeDriver::close()` starts by calling `->reset()` which
internally closes all non-main windows. The subsequent attempt
to enumerate windows and close all but the first is therefore
redundant.

It is possible this was historically helping with the timing
issue of windows closing asynchronously after reset(), however
this has now been resolved and the windows are guaranteed to
be closed by the time reset() returns.