ci: Configure phpunit with `<env` rather than `<server` vars
The difference is that `<server>` will *override* environment variables
that are already set, where `<env>` provides a default for use if the
var is not defined.

This allows us to use a single phpunit configuration in most
environments (including local dev and CI) and set the CHROME_URL and
WEB_FIXTURES_URL from environment variables as expected.

Note that we now have to override getWebFixturesUrl in the
ChromeDriverConfig because the default in mink/driver-testsuite only
reads from $_SERVER.